+++ CHANGELOG.md
@@ -1,7 +1,27 @@
 # Changelog
+
 All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+### Added
+
+- Project setup.
+- Essential dependencies for interface styling.
+- Basic interface.
+- A cool logo.
+- A `cls` method to clear the command line interface.
+- `Player` class.
+- `Game` class.
+- `Engine` class.
+- Take inputs interchangebly from players.
+- defining row, column and diagonal? of last move in `Engine` class.
+- The game logic.
+- Game Instructions readme file.
+- Test the engine class.
+- Test the player class.
+- Test the game class.
+- Game design refactoring.